Fix column name error in node query
Updated the query to ensure it references the correct column names in the Supabase database, resolving the "column node.x does not exist" error.
[skip gpt_engineer]

+++ supabase/migrations/20240325000000_initial_schema.sql
@@ -14,8 +14,8 @@ CREATE TABLE public.node (
     owner_id UUID REFERENCES public.profiles(id),
     investigation_id UUID REFERENCES public.investigations(id),
     parent_node_id UUID REFERENCES public.node(id) ON DELETE SET NULL,
-    x NUMERIC(10,2) DEFAULT 0,
-    y NUMERIC(10,2) DEFAULT 0,
+    x DOUBLE PRECISION DEFAULT 0,
+    y DOUBLE PRECISION DEFAULT 0,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
